move images to /public and add img filter to images on landingHome
