+++ wwwroot/src/pages/WhyUs.js
@@ -5,8 +5,8 @@ import { FeatureCollection } from './Landing.js';
 import Layout from '../components/Layout';
 import { makeStyles } from '@material-ui/core/styles';
 
-import bannerImg from '../img/why-us/banner.jpg';
-import gettingImg from '../img/why-us/getting-better.jpg';
+import bannerImg from '../../public/whyus-banner.jpg';
+import gettingImg from '../../public/whyus-gettingbetter.jpg';
 
 const useStyles = makeStyles((theme) => ({
   root: {}
